IEEE-205 rename global_css to sharedStyles and add lower case number type
