+++ hackathon_site/dashboard/frontend/src/components/sharedStyles/Filter.module.scss
